+++ astropy/io/votable/tests/vo_test.py
@@ -718,7 +718,8 @@ def table_from_scratch():
 
 def test_open_files():
     for filename in get_pkg_data_filenames('data', pattern='*.xml'):
-        if filename.endswith('custom_datatype.xml'):
+        if (filename.endswith('custom_datatype.xml') or
+                filename.endswith('timesys_errors.xml')):
             continue
         parse(filename)
 
@@ -792,7 +793,7 @@ def test_validate(test_path_object=False):
 
     # Uncomment to generate new groundtruth
     # with open('validation.txt', 'wt', encoding='utf-8') as fd:
-    #     fd.write(u''.join(output))
+    #    fd.write(u''.join(output))
 
     with open(
         get_pkg_data_filename('data/validation.txt'),
@@ -997,3 +998,52 @@ def test_custom_datatype():
 
     table = votable.get_first_table()
     assert table.array.dtype['foo'] == np.int32
+
+
+def _timesys_tests(votable):
+    assert len(list(votable.iter_timesys())) == 4
+
+    timesys = votable.get_timesys_by_id('time_frame')
+    assert timesys.timeorigin == 2455197.5
+    assert timesys.timescale == 'TCB'
+    assert timesys.refposition == 'BARYCENTER'
+
+    timesys = votable.get_timesys_by_id('mjd_origin')
+    assert timesys.timeorigin == 'MJD-origin'
+    assert timesys.timescale == 'TDB'
+    assert timesys.refposition == 'EMBARYCENTER'
+
+    timesys = votable.get_timesys_by_id('jd_origin')
+    assert timesys.timeorigin == 'JD-origin'
+    assert timesys.timescale == 'TT'
+    assert timesys.refposition == 'HELIOCENTER'
+
+    timesys = votable.get_timesys_by_id('no_origin')
+    assert timesys.timeorigin is None
+    assert timesys.timescale == 'UTC'
+    assert timesys.refposition == 'TOPOCENTER'
+
+
+def test_timesys():
+    votable = parse(get_pkg_data_filename('data/timesys.xml'))
+    _timesys_tests(votable)
+
+
+def test_timesys_roundtrip():
+    orig_votable = parse(get_pkg_data_filename('data/timesys.xml'))
+    bio = io.BytesIO()
+    orig_votable.to_xml(bio)
+    bio.seek(0)
+    votable = parse(bio)
+    _timesys_tests(votable)
+
+
+def test_timesys_errors():
+    output = io.StringIO()
+    validate(get_pkg_data_filename('data/timesys_errors.xml'), output,
+             xmllint=False)
+    outstr = output.getvalue()
+    assert("E23: Invalid timeorigin attribute 'bad-origin'" in outstr)
+    assert("E22: ID attribute is required for all TIMESYS elements" in outstr)
+    assert("W48: Unknown attribute 'refposition_mispelled' on TIMESYS"
+           in outstr)